added some training code!!
